requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
PyPDF2==3.0.1
python-docx==1.1.2
pdfminer.six==20240706
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0
python-magic==0.4.27
celery==5.4.0
redis==5.0.8
gunicorn==22.0.0
htmx==1.9.10
